Dictionary codec builders

+++ modules/core/src/main/scala/io/taig/otter/Constraint.scala
@@ -20,8 +20,8 @@ object Constraint:
     case UniqueItems
 
   enum Object:
-    case MaxProperties(reference: Int)
-    case MinProperties(reference: Int)
+    case MaxProperties(reference: Long)
+    case MinProperties(reference: Long)
 
   enum Primitive[+A]:
     case Matches(pattern: Pattern)
